+++ ChangeLog.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 03/16/2020 0.21.1
+
+- Enabled request parameter suppression in model validation.
+
+## 03/16/2020 0.21.0
+
+- Fixed oav does not report the invalid internal references error issue. Added a function (verfiyInternalReference) to check the error.
+
 ## 03/13/2020 0.20.13
 
 - Normalize the path parameter values by using whilelist (0-9a-zA-Z._) to fix the issue of validate-example failed